# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Payment Gateway API",
    "version": "1.0.0",
    "description": "Simple payment gateway API with Midtrans integration"
  },
    "servers": [
    {
      "url": "https://mansur.my.id",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000", 
      "description": "Local Development server"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": { "type": "string" }
              }
            }
          }
        }
      }
    },
    "/invoices": {
      "post": {
        "summary": "Create invoice",
        "security": [{"ApiKeyAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {"type": "integer"}
                },
                "required": ["amount"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invoice created",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Invoice"}
              }
            }
          }
        }
      }
    },
    "/invoices/{id}": {
      "get": {
        "summary": "Get invoice (legacy)",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {
          "200": {
            "description": "Invoice details",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Invoice"}
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete invoice",
        "security": [{"ApiKeyAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {
          "200": {
            "description": "Invoice deleted",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Invoice"}
              }
            }
          }
        }
      }
    },
    "/v1/invoices": {
      "get": {
        "summary": "List all invoices",
        "security": [{"ApiKeyAuth": []}],
        "responses": {
          "200": {
            "description": "List of invoices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {"$ref": "#/components/schemas/Invoice"}
                }
              }
            }
          }
        }
      }
    },
    "/v1/payment-links": {
      "post": {
        "summary": "Generate payment link",
        "security": [{"ApiKeyAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "order_id": {"type": "string"}
                },
                "required": ["order_id"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment link generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "payment_url": {"type": "string"},
                    "order_id": {"type": "string"},
                    "payment_id": {"type": "string"},
                    "expires_in": {"type": "integer"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/midtrans/notification": {
      "post": {
        "summary": "Midtrans webhook",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "transaction_time": {"type": "string"},
                  "transaction_status": {"type": "string", "enum": ["pending", "settlement", "cancel", "deny", "expire", "failure"]},
                  "transaction_id": {"type": "string"},
                  "status_message": {"type": "string"},
                  "status_code": {"type": "string"},
                  "signature_key": {"type": "string"},
                  "payment_type": {"type": "string"},
                  "order_id": {"type": "string"},
                  "merchant_id": {"type": "string"},
                  "gross_amount": {"type": "string"},
                  "fraud_status": {"type": "string"},
                  "currency": {"type": "string"}
                },
                "required": ["order_id", "transaction_status", "gross_amount"]
              },
              "example": {
                "transaction_time": "2024-01-15 10:30:00",
                "transaction_status": "settlement",
                "transaction_id": "b3a4d5e6-7890-1234-5678-90abcdef1234",
                "status_message": "midtrans payment notification",
                "status_code": "200",
                "signature_key": "abc123def456...",
                "payment_type": "bank_transfer",
                "order_id": "ORDER-123456789",
                "merchant_id": "G123456789",
                "gross_amount": "100000.00",
                "fraud_status": "accept",
                "currency": "IDR"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {"type": "string"}
                  }
                },
                "example": {
                  "status": "OK"
                }
              }
            }
          }
        }
      }
    },

  "components": {
    "schemas": {
      "Invoice": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "orderId": {"type": "string"},
          "amount": {"type": "integer"},
          "status": {"type": "string", "enum": ["PENDING", "PAID", "FAILED", "EXPIRED"]},
          "createdAt": {"type": "string", "format": "date-time"},
          "updatedAt": {"type": "string", "format": "date-time"},
          "paymentLink": {"type": "string", "nullable": true},
          "paymentLinkCreatedAt": {"type": "string", "format": "date-time", "nullable": true}
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {"type": "string"}
        }
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  }
  }
}